Fixed extension of engine model.
Extending or reopening a model from an engine within an app works fine
with cache_classes = true. When class reloading is turned on (in
development), the model to be extended does not get reloaded. It looks
like `to_prepare` is the way to go in this case. It ensures that the
wrapped code gets executed before every request with class reloading
turned on.